Implement Command Pattern in BadBM to further decouple benchmark
