+++ src/main/java/BadBM/DiskWorker.java
@@ -1,21 +1,11 @@
 package BadBM;
 
-import BadBM.persist.DiskRun;
-import BadBM.persist.EM;
-import BadBM.ui.Gui;
+import BadBM.commands.ReadBenchmarkOperation;
+import BadBM.commands.WriteBenchmarkOperation;
 
-import javax.persistence.EntityManager;
-import java.io.File;
-import java.io.FileNotFoundException;
 import java.io.IOException;
-import java.io.RandomAccessFile;
-import java.util.Date;
-import java.util.logging.Level;
-import java.util.logging.Logger;
 
-import static BadBM.App.*;
-import static BadBM.DiskMark.MarkType.READ;
-import static BadBM.DiskMark.MarkType.WRITE;
+import static BadBM.App.msg;
 
 /**
  * Run the disk benchmark. Updates the ProgramInterface Object to allow either GUI or console to show run data
@@ -32,7 +22,7 @@ import static BadBM.DiskMark.MarkType.WRITE;
 
 public class DiskWorker extends Thread {
 
-    private ProgramInterface programInterface;
+    private final ProgramInterface programInterface;
 
     public DiskWorker(ProgramInterface userInterface) {
         programInterface = userInterface;
@@ -56,127 +46,23 @@ public class DiskWorker extends Thread {
     public Boolean doBenchmark() throws IOException {
 
         programInterface.setRunning(true);
+        int unitsComplete = 0;
         System.out.println("*** starting new worker thread");
         msg("Running readTest " + App.readTest + "   writeTest " + App.writeTest);
         msg("num files: " + App.numOfMarks + ", num blks: " + App.numOfBlocks
                 + ", blk size (kb): " + App.blockSizeKb + ", blockSequence: " + App.blockSequence);
 
-        /**
-         * init local vars that keep track of benchmarks, and a large read/write buffer
-         */
-        int wUnitsComplete = 0, rUnitsComplete = 0, unitsComplete;
-        int wUnitsTotal = App.writeTest ? numOfBlocks * numOfMarks : 0;
-        int rUnitsTotal = App.readTest ? numOfBlocks * numOfMarks : 0;
-        int unitsTotal = wUnitsTotal + rUnitsTotal;
-        float percentComplete;
-
-        int blockSize = blockSizeKb * KILOBYTE;
-        byte[] blockArr = new byte[blockSize];
-        for (int b = 0; b < blockArr.length; b++) {
-            if (b % 2 == 0) {
-                blockArr[b] = (byte) 0xFF;
-            }
-        }
-        
-        DiskMark wMark, rMark;  // declare vars that will point to objects used to pass progress to UI
-
-        int startFileNum = App.nextMarkNumber;
+        BenchmarkExecutor executor = new BenchmarkExecutor();
 
         if (App.writeTest) {
-            DiskRun run = initializeRun(DiskRun.IOMode.WRITE);
-
-            programInterface.setTitle(run.getDiskInfo());
-
-            createDataFile();
-
-            /**
-             * Begin an outer loop for specified duration (number of 'marks') of benchmark,
-             * that keeps writing data (in its own loop - for specified # of blocks). Each 'Mark' is timed
-             * and is reported to the GUI for display as each Mark completes.
-             */
-            for (int m = startFileNum; m < startFileNum + App.numOfMarks && !programInterface.isCancel(); m++) {
-
-                if (App.multiFile) {
-                    testFile = new File(dataDir.getAbsolutePath()
-                            + File.separator + "testdata" + m + ".jdm");
-                }
-                wMark = new DiskMark(WRITE);    // starting to keep track of a new bench Mark
-                wMark.setMarkNum(m);
-                long startTime = System.nanoTime();
-                long totalBytesWrittenInMark = 0;
-
-                String mode = "rw";
-                if (App.writeSyncEnable) {
-                    mode = "rwd";
-                }
-
-                try {
-                    try (RandomAccessFile rAccFile = new RandomAccessFile(testFile, mode)) {
-                        for (int b = 0; b < numOfBlocks; b++) {
-                            if (App.blockSequence == DiskRun.BlockSequence.RANDOM) {
-                                int rLoc = Util.randInt(0, numOfBlocks - 1);
-                                rAccFile.seek(rLoc * blockSize);
-                            } else {
-                                rAccFile.seek(b * blockSize);
-                            }
-                            rAccFile.write(blockArr, 0, blockSize);
-                            totalBytesWrittenInMark += blockSize;
-                            wUnitsComplete++;
-                            unitsComplete = rUnitsComplete + wUnitsComplete;
-                            percentComplete = (float) unitsComplete / (float) unitsTotal * 100f;
-
-                            /**
-                             * Report to GUI what percentage level of Entire BM (#Marks * #Blocks) is done.
-                             */
-                            programInterface.setProgressVal((int) percentComplete);
-                        }
-                    }
-                } catch (IOException ex) {
-                    Logger.getLogger(App.class.getName()).log(Level.SEVERE, null, ex);
-                }
-
-                /**
-                 * Compute duration, throughput of this Mark's step of BM
-                 */
-                long endTime = System.nanoTime();
-                long elapsedTimeNs = endTime - startTime;
-                double sec = (double) elapsedTimeNs / (double) 1000000000;
-                double mbWritten = (double) totalBytesWrittenInMark / (double) MEGABYTE;
-                wMark.setBwMbSec(mbWritten / sec);
-                msg("m:" + m + " write IO is " + wMark.getBwMbSecAsString() + " MB/s     "
-                        + "(" + Util.displayString(mbWritten) + "MB written in "
-                        + Util.displayString(sec) + " sec)");
-                App.updateMetrics(wMark);
-
-                /**
-                 * Let the GUI know the interim result described by the current Mark
-                 */
-                programInterface.setMark(wMark);
-
-                // Keep track of statistics to be displayed and persisted after all Marks are done.
-                run.setRunMax(wMark.getCumMax());
-                run.setRunMin(wMark.getCumMin());
-                run.setRunAvg(wMark.getCumAvg());
-                run.setEndTime(new Date());
-            } // END outer loop for specified duration (number of 'marks') for WRITE bench mark
-
-            /**
-             * Persist info about the Write BM Run (e.g. into Derby Database) and add it to a GUI panel
-             */
-            EntityManager em = EM.getEntityManager();
-            em.getTransaction().begin();
-            em.persist(run);
-            em.getTransaction().commit();
-
-            Gui.runPanel.addRun(run);
-
+            unitsComplete = executor.executeBenchmarkOperation(new WriteBenchmarkOperation(programInterface, unitsComplete));
         }
-        /**
-         * Most benchmarking systems will try to do some cleanup in between 2 benchmark operations to
-         * make it more 'fair'. For example a networking benchmark might close and re-open sockets,
-         * a memory benchmark might clear or invalidate the Op Systems TLB or other caches, etc.
-         */
 
+        /*
+          Most benchmarking systems will try to do some cleanup in between 2 benchmark operations to
+          make it more 'fair'. For example a networking benchmark might close and re-open sockets,
+          a memory benchmark might clear or invalidate the Op Systems TLB or other caches, etc.
+         */
         // try renaming all files to clear catch
         if (App.readTest && App.writeTest && !programInterface.isCancel()) {
             programInterface.displayMessage("For valid READ measurements please clear the disk cache by\n" +
@@ -189,92 +75,12 @@ public class DiskWorker extends Thread {
 
         // Same as above, just for Read operations instead of Writes.
         if (App.readTest) {
-            DiskRun run = initializeRun(DiskRun.IOMode.READ);
-
-            Gui.chartPanel.getChart().getTitle().setVisible(true);
-            Gui.chartPanel.getChart().getTitle().setText(run.getDiskInfo());
-
-            for (int m = startFileNum; m < startFileNum + App.numOfMarks && !programInterface.isCancel(); m++) {
-
-                if (App.multiFile) {
-                    testFile = new File(dataDir.getAbsolutePath()
-                            + File.separator + "testdata" + m + ".jdm");
-                }
-                rMark = new DiskMark(READ);
-                rMark.setMarkNum(m);
-                long startTime = System.nanoTime();
-                long totalBytesReadInMark = 0;
-
-                try {
-                    try (RandomAccessFile rAccFile = new RandomAccessFile(testFile, "r")) {
-                        for (int b = 0; b < numOfBlocks; b++) {
-                            if (App.blockSequence == DiskRun.BlockSequence.RANDOM) {
-                                int rLoc = Util.randInt(0, numOfBlocks - 1);
-                                rAccFile.seek(rLoc * blockSize);
-                            } else {
-                                rAccFile.seek(b * blockSize);
-                            }
-                            rAccFile.readFully(blockArr, 0, blockSize);
-                            totalBytesReadInMark += blockSize;
-                            rUnitsComplete++;
-                            unitsComplete = rUnitsComplete + wUnitsComplete;
-                            percentComplete = (float) unitsComplete / (float) unitsTotal * 100f;
-                            programInterface.setProgressVal((int) percentComplete);
-                        }
-                    }
-                } catch (FileNotFoundException ex) {
-                    Logger.getLogger(App.class.getName()).log(Level.SEVERE, null, ex);
-                }
-                long endTime = System.nanoTime();
-                long elapsedTimeNs = endTime - startTime;
-                double sec = (double) elapsedTimeNs / (double) 1000000000;
-                double mbRead = (double) totalBytesReadInMark / (double) MEGABYTE;
-                rMark.setBwMbSec(mbRead / sec);
-                msg("m:" + m + " READ IO is " + rMark.getBwMbSec() + " MB/s    "
-                        + "(MBread " + mbRead + " in " + sec + " sec)");
-                App.updateMetrics(rMark);
-                programInterface.setMark(rMark);
-
-                run.setRunMax(rMark.getCumMax());
-                run.setRunMin(rMark.getCumMin());
-                run.setRunAvg(rMark.getCumAvg());
-                run.setEndTime(new Date());
-            }
-
-            EntityManager em = EM.getEntityManager();
-            em.getTransaction().begin();
-            em.persist(run);
-            em.getTransaction().commit();
-
-            Gui.runPanel.addRun(run);
+            unitsComplete = executor.executeBenchmarkOperation(new ReadBenchmarkOperation(programInterface, unitsComplete));
         }
+
         App.nextMarkNumber += App.numOfMarks;
         programInterface.setRunning(false);
         programInterface.setFinished(true); //needed for console test
         return true;
-    }
-
-    private void createDataFile() {
-        // Create a test data file using the default file system and config-specified location
-        if (!App.multiFile) {
-            testFile = new File(dataDir.getAbsolutePath() + File.separator + "testdata.jdm");
-        }
-    }
-
-    private DiskRun initializeRun(DiskRun.IOMode write) {
-        DiskRun run = new DiskRun(write, App.blockSequence);
-        setRunConfig(run);
-
-        // Tell logger and GUI to display what we know so far about the Run
-        msg("disk info: (" + run.getDiskInfo() + ")");
-        return run;
-    }
-
-    private void setRunConfig(DiskRun run) {
-        run.setNumMarks(App.numOfMarks);
-        run.setNumBlocks(App.numOfBlocks);
-        run.setBlockSize(App.blockSizeKb);
-        run.setTxSize(App.targetTxSizeKb());
-        run.setDiskInfo(Util.getDiskInfo(dataDir));
     }
 }
