+++ src/main/java/BadBM/commands/BenchmarkOperation.java
@@ -1,0 +1,5 @@
+package BadBM.commands;
+
+public interface BenchmarkOperation {
+    int execute();
+}
