+++ src/main/java/BadBM/commands/ReadBenchmark.java
@@ -1,0 +1,137 @@
+package BadBM.commands;
+
+import BadBM.App;
+import BadBM.DiskMark;
+import BadBM.ProgramInterface;
+import BadBM.Util;
+import BadBM.persist.DiskRun;
+import BadBM.persist.EM;
+import BadBM.ui.Gui;
+
+import javax.persistence.EntityManager;
+import java.io.File;
+import java.io.FileNotFoundException;
+import java.io.IOException;
+import java.io.RandomAccessFile;
+import java.util.Date;
+import java.util.logging.Level;
+import java.util.logging.Logger;
+
+import static BadBM.App.*;
+import static BadBM.App.msg;
+import static BadBM.DiskMark.MarkType.READ;
+
+/**
+ * Receiver class for Read Benchmark commands. Performs the actual benchmark operation. The DiskRun class is used to
+ * keep track of and persist info about each benchmark at a higher level (a run), while the DiskMark class describes
+ * each iteration's result, which is displayed by the UI as the benchmark run progresses.
+ */
+public class ReadBenchmark {
+
+    public int doReadBenchmark(ProgramInterface programInterface, int units) {
+        // declare local vars formerly in DiskWorker
+
+        int wUnitsComplete = units,
+            rUnitsComplete = 0,
+            unitsComplete;
+
+        int wUnitsTotal = App.writeTest ? numOfBlocks * numOfMarks : 0;
+        int rUnitsTotal = App.readTest ? numOfBlocks * numOfMarks : 0;
+        int unitsTotal = wUnitsTotal + rUnitsTotal;
+        float percentComplete;
+
+        int blockSize = blockSizeKb*KILOBYTE;
+        byte [] blockArr = new byte [blockSize];
+        for(int b=0; b < blockArr.length; b++) {
+            if (b%2==0) {
+                blockArr[b]=(byte)0xFF;
+            }
+        }
+
+        DiskMark rMark;
+        int startFileNum = App.nextMarkNumber;
+
+        DiskRun run = initializeRun(DiskRun.IOMode.READ);
+
+        Gui.chartPanel.getChart().getTitle().setVisible(true);
+        Gui.chartPanel.getChart().getTitle().setText(run.getDiskInfo());
+
+        for (int m = startFileNum; m < startFileNum + App.numOfMarks && !programInterface.isCancel(); m++) {
+
+            if (App.multiFile) {
+                testFile = new File(dataDir.getAbsolutePath()
+                        + File.separator + "testdata" + m + ".jdm");
+            }
+            rMark = new DiskMark(READ);
+            rMark.setMarkNum(m);
+            long startTime = System.nanoTime();
+            long totalBytesReadInMark = 0;
+
+            try (RandomAccessFile rAccFile = new RandomAccessFile(testFile, "r")) {
+                for (int b = 0; b < numOfBlocks; b++) {
+                    if (App.blockSequence == DiskRun.BlockSequence.RANDOM) {
+                        int rLoc = Util.randInt(0, numOfBlocks - 1);
+                        rAccFile.seek(rLoc * blockSize);
+                    } else {
+                        rAccFile.seek(b * blockSize);
+                    }
+                    rAccFile.readFully(blockArr, 0, blockSize);
+                    totalBytesReadInMark += blockSize;
+                    rUnitsComplete++;
+                    unitsComplete = rUnitsComplete + wUnitsComplete;
+                    percentComplete = (float) unitsComplete / (float) unitsTotal * 100f;
+                    programInterface.setProgressVal((int) percentComplete);
+                }
+            } catch (IOException e) {
+                e.printStackTrace();
+            }
+            long endTime = System.nanoTime();
+            long elapsedTimeNs = endTime - startTime;
+            double sec = (double) elapsedTimeNs / (double) 1000000000;
+            double mbRead = (double) totalBytesReadInMark / (double) MEGABYTE;
+            rMark.setBwMbSec(mbRead / sec);
+            msg("m:" + m + " READ IO is " + rMark.getBwMbSec() + " MB/s    "
+                    + "(MBread " + mbRead + " in " + sec + " sec)");
+            App.updateMetrics(rMark);
+            programInterface.setMark(rMark);
+
+            run.setRunMax(rMark.getCumMax());
+            run.setRunMin(rMark.getCumMin());
+            run.setRunAvg(rMark.getCumAvg());
+            run.setEndTime(new Date());
+        }
+
+        EntityManager em = EM.getEntityManager();
+        em.getTransaction().begin();
+        em.persist(run);
+        em.getTransaction().commit();
+
+        Gui.runPanel.addRun(run);
+        return rUnitsComplete;
+    }
+
+    private void createDataFile() {
+        // Create a test data file using the default file system and config-specified location
+        if (!App.multiFile) {
+            testFile = new File(dataDir.getAbsolutePath() + File.separator + "testdata.jdm");
+        }
+    }
+
+    private DiskRun initializeRun(DiskRun.IOMode write) {
+        DiskRun run = new DiskRun(write, App.blockSequence);
+        setRunConfig(run);
+
+        // Tell logger and GUI to display what we know so far about the Run
+        msg("disk info: (" + run.getDiskInfo() + ")");
+        return run;
+    }
+
+    private void setRunConfig(DiskRun run) {
+        run.setNumMarks(App.numOfMarks);
+        run.setNumBlocks(App.numOfBlocks);
+        run.setBlockSize(App.blockSizeKb);
+        run.setTxSize(App.targetTxSizeKb());
+        run.setDiskInfo(Util.getDiskInfo(dataDir));
+    }
+
+}
